Prototype an Ibexa Cloud installation script

+++ install-ibexa-cloud.bash
@@ -6,6 +6,7 @@
 set -e;
 set -x;
 
+
 IBEXA_DXP_VERSION=$1; # content, experience, commerce, experience:v3.3.2, commerce:^3.3@rc, …
 IBEXA_REPO_USER=$2; # a.k.a <installation-key>
 IBEXA_REPO_PASS=$3; # a.k.a <token-password>
@@ -17,11 +18,43 @@ PLATFORM_PROJECT_NAME=$8;
 PLATFORM_TOKEN=$9;
 DEFAULT_BRANCH=${10:-'master'}; # production, main, staging, …
 
-platform integration:add --yes --project=$PLATFORM_PROJECT_ID --type=github --token=$GITHUB_TOKEN --repository=$GITHUB_REPO --build-pull-requests=false --fetch-branches=true --prune-branches=true ;
+EMPTY_LOCAL_REPO_DIR=1;
+GIT_IGNORE_AUTH_JSON=1;
+FILL_PSH_CONTENT_FIRST=0;
+
+
+function wait_platform_activity()
+{
+  if [ -n "$1" ]; then
+    project_option="--project=$1";
+  else
+    project_option='';
+  fi;
+  if [ -n "$2" ]; then
+    environment_option="--environment=$2";
+  else
+    environment_option='';
+  fi;
+  sleep 20;
+  while [ -n "$(platform activity:list --incomplete $project_option $environment_option --no-header --format=plain 2> /dev/null)" ]; do
+    sleep 10;
+  done
+}
+
+function sedi() {
+  ## Regular
+  #sed -i -e $*;
+  ## MacOS
+  sed -i '' -e "$1" $2;
+}
+
+
+platform integration:add --yes --project=$PLATFORM_PROJECT_ID --type=github --token=$GITHUB_TOKEN --repository=$GITHUB_REPO --build-pull-requests=false --build-draft-pull-requests=false --fetch-branches=true --prune-branches=true;
 integration=`platform integration:list --project=$PLATFORM_PROJECT_ID --format=csv | grep $GITHUB_REPO;`;
 platform integration:validate --project=$PLATFORM_PROJECT_ID ${integration%%,*};
+platform integration:get --project=$PLATFORM_PROJECT_ID ${integration%%,*};
 
-if [ 1 -eq 1 ]; then
+if [ 1 -eq $EMPTY_LOCAL_REPO_DIR ]; then
   cd && rm -rf $LOCAL_REPO_DIR;
 fi;
 if [ -e $LOCAL_REPO_DIR ]; then
@@ -38,14 +71,16 @@ fi;
 composer create-project ibexa/$skeleton ./ --no-install;
 rm -v composer.lock;
 composer config platform.php 7.3.23;
-composer config http-basic.updates.ibexa.co $IBEXA_REPO_USER $IBEXA_REPO_PASS;
-composer config repositories.ibexa composer https://updates.ibexa.co;
+composer config http-basic.updates.ibexa.co $IBEXA_REPO_USER $IBEXA_REPO_PASS; # auth.json
+#composer config repositories.ibexa composer https://updates.ibexa.co;
 composer install --no-scripts;
 
 composer ibexa:setup --platformsh;
 
-cp .gitignore.dist .gitignore;
-if [ 1 -eq 1 ]; then
+if [ ! -f .gitignore ] && [ -f .gitignore.dist ]; then
+  cp .gitignore.dist .gitignore;
+fi;
+if [ 1 -eq $GIT_IGNORE_AUTH_JSON ]; then
   echo 'auth.json' >> .gitignore;
   platform variable:create --project=$PLATFORM_PROJECT_ID --level=project --name=env:COMPOSER_AUTH \
     --json=true --visible-runtime=false --sensitive=true --visible-build=true \
@@ -62,12 +97,12 @@ git commit -m "create-project ibexa/$skeleton ./";
 git branch -M $DEFAULT_BRANCH;
 #git push -u origin $DEFAULT_BRANCH;
 git push -u platform $DEFAULT_BRANCH;
+wait_platform_activity $PLATFORM_PROJECT_ID;
 
-sleep 60;
 if [ 'master' != "$DEFAULT_BRANCH" ]; then
   platform project:info default_branch $DEFAULT_BRANCH --project=$PLATFORM_PROJECT_ID;
+  wait_platform_activity $PLATFORM_PROJECT_ID;
   platform environment:info --project=$PLATFORM_PROJECT_ID --environment=$DEFAULT_BRANCH parent -;
-  sleep 30;
   platform environment:delete master --delete-branch --project=$PLATFORM_PROJECT_ID --yes;
 fi;
 
@@ -76,22 +111,24 @@ if [ '' != "$(git diff)" ]; then
   git add ./;
   git commit -m "recipes:install ibexa/${IBEXA_DXP_VERSION%:*}";
   git push;
+  wait_platform_activity $PLATFORM_PROJECT_ID $DEFAULT_BRANCH;
 fi;
 
-sed -i '' 's/10.15.3/10.19.0/' .platform.app.yaml;
+sedi 's/10.15.3/10.19.0/' .platform.app.yaml;
 git add ./;
 git commit -m 'Update node version
 
 error @symfony/webpack-encore@1.1.2: The engine "node" is incompatible with this module. Expected version "^10.19.0 || ^12.0.0 || >=14.0.0". Got "10.15.3"';
 git push;
+wait_platform_activity $PLATFORM_PROJECT_ID $DEFAULT_BRANCH;
 
-sleep 60; #300; #600;
-platform environment:activate --yes $DEFAULT_BRANCH;
+platform environment:activate --yes --no-wait $DEFAULT_BRANCH;
+wait_platform_activity $PLATFORM_PROJECT_ID $DEFAULT_BRANCH;
 
 lando init --source cwd --recipe platformsh --platformsh-site "$PLATFORM_PROJECT_NAME" --platformsh-auth $PLATFORM_TOKEN;
-lando start;
+lando rebuild;
 
-if [ 1 -eq 1 ]; then
+if [ 1 -eq $FILL_PSH_CONTENT_FIRST ]; then
   echo "Populate P.sh current project and environment database";
   platform environment:ssh php bin/console ibexa:install;
   echo "Download/Copy content to local Lando";
@@ -112,7 +149,7 @@ platform environment:ssh php bin/console ibexa:graphql:generate-schema;
 lando php bin/console ibexa:graphql:generate-schema;
 
 
-sed -i '' 's/        composer install --no-dev --prefer-dist --no-progress --no-interaction --optimize-autoloader/        if [ "dev" = "$APP_ENV" ]; then\
+sedi 's/        composer install --no-dev --prefer-dist --no-progress --no-interaction --optimize-autoloader/        if [ "dev" = "$APP_ENV" ]; then\
             composer install --dev --no-interaction\
         else\
             composer install --no-dev --prefer-dist --no-progress --no-interaction --optimize-autoloader\
@@ -121,6 +158,8 @@ sed -i '' 's/        composer install --no-dev --prefer-dist --no-progress --no-
 git add .platform.app.yaml;
 git commit -m '.platform.app.yaml: Add APP_ENV=dev possibility';
 git push;
+wait_platform_activity $PLATFORM_PROJECT_ID $DEFAULT_BRANCH;
+
 
 # https://docs.lando.dev/config/platformsh.html#environment-variables
 echo "  overrides:
@@ -128,21 +167,23 @@ echo "  overrides:
       variables:
         env:
           APP_ENV: dev" >> .lando.yml;
-lando start;
+lando rebuild;
 
 
 vendor/ezsystems/ezplatform-solr-search-engine/bin/generate-solr-config.sh;
-sed -i '' 's/#\(solr:.*\)/\1/' .platform.app.yaml;
-sed -i '' '/solrsearch/,/core: collection1/ s/^#//' .platform/services.yaml;
+sedi 's/#\(solr:.*\)/\1/' .platform.app.yaml;
+sedi '/solrsearch/,/core: collection1/ s/^#//' .platform/services.yaml;
 git add .platform/configsets/ .platform.app.yaml .platform/services.yaml;
 git commit -m 'Enable Solr';
 git push;
+wait_platform_activity $PLATFORM_PROJECT_ID $DEFAULT_BRANCH;
+
 
 # Avoid having 'http://solr.internal:8080/solr/solrsearch'
 #echo '          SOLR_CORE: collection1' >> .lando.yml;
 #echo '          SISO_SEARCH_SOLR_CORE: collection1' >> .lando.yml;
-#lando start;
-sed -i '' 's/%solr_core%/collection1/' config/packages/ezplatform_solr.yaml;
+#lando rebuild;
+sedi 's/%solr_core%/collection1/' config/packages/ezplatform_solr.yaml;
 lando php bin/console cache:clear;
 lando php bin/console ibexa:reindex;
